Refactor auth imports to use types/auth module
Updated imports in bearerTokenProvider.ts and oAuthCredentials.ts to reference types and schemas from the new types/auth module instead of zuoraSchemas. This improves code organization and modularity.
